requirements.txt
Flask
mysql-connector-python
pandas
flask_mysqldb
